binder: support dedicated node threads
This change allows a node to be bound to a certain thread,
meaning that all transactions to the node will be handled
by that specific thread. Such threads won't be dealing
with generic process work.

This requires an additional todo list on the thread to
hold work, as multiple nodes may be assigned to the same
thread.

To introduce this change in a way that is backwards
compatible with older userspaces, this change adds the
flat_binder_object_thread structure, which is identical
to flat_binder_object, except that it has a thread field
added at the end. When FLAT_BINDER_FLAG_THREAD is set,
in the flags field of flat_binder_object, the kernel
knows that userspace is using the new version of the
structure.

Change-Id: I4e252eca4e00c337de859600bc6caf5131120599
Signed-off-by: Martijn Coenen <maco@android.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>